fix: use git hash for upgrade test
In upgrade test, use the hash of the original branch to switch back after checkout of master/main. This fixes an issue where the original PR branch is a detached HEAD reference, and does not have a suitable "branch name" to checkout at end of test, but the hash will always work.