fix: use git hash for upgrade test checkouts (#109)
* fix: remove fetching of HEAD for upgrade test

Removing the fetch of HEAD, as it is also creating a local branch called HEAD which will cause git to complain about ambiguous HEAD references (HEAD as a "current branch pointer" and a local branch name). In testing was found it was not necessary for functionality.

* fix: use git hash for upgrade test

In upgrade test, use the hash of the original branch to switch back after checkout of master/main. This fixes an issue where the original PR branch is a detached HEAD reference, and does not have a suitable "branch name" to checkout at end of test, but the hash will always work.